Reorganize navigation state integration test files
